fix: handle beginning/end of buffer case for before/after pair functions.
Without doing this, problems can happen in some corner cases. e.g. A region is selected, the point is at beginning of buffer, the user want to press DEL to delete region, while DEL is bound to `awesome-pair-backward-delete`. Then `awesome-pair-after-close-pair-p` will signal an error and not do anything. So let's handle this and prevent similar problems once and for all.